requirements.txt
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.1
python-decouple==3.3
Werkzeug==0.16.1
pytest==5.4.1
